all routes are done now working on email and then finishing the FE connection to the backend

+++ backend/src/routers/notes.js
@@ -68,7 +68,7 @@ router.patch ("/Notes/:id", auth, async (req, res) => {
 
 router.delete("/Notes/:id", auth, async (req, res) => {
   try{
-    const note = await Note.findByIdAndDelete(req.params.id)
+    const note = await Note.findByIdAndDelete({ _id: req.params.id, owner: req.user._id}); 
 
     if (!note) {
       return res.status(404).send()
